+++ .rustfmt.toml
@@ -7,5 +7,5 @@ reorder_imports = true
 # format_strings = false
 # comment_width = 100
 # wrap_comments = true
-# group_imports = "StdExternalCrate"
-# imports_granularity = "Module"
+group_imports = "StdExternalCrate"
+imports_granularity = "Module"
