+++ crates/cli/cli-components/src/traits/config/config_path.rs
@@ -1,5 +1,4 @@
 use cgp_core::prelude::*;
-
 use hermes_runtime_components::traits::fs::file_path::{FilePathOf, HasFilePathType};
 use hermes_runtime_components::traits::runtime::HasRuntime;
