+++ crates/cosmos/cosmos-integration-tests/src/contexts/binary_channel/test_driver.rs
@@ -6,17 +6,17 @@ use hermes_logger::ProvideHermesLogger;
 use hermes_logging_components::traits::has_logger::{
     GlobalLoggerGetterComponent, LoggerGetterComponent, LoggerTypeComponent,
 };
+use hermes_relayer_components::multi::traits::birelay_at::BiRelayTypeAtComponent;
+use hermes_relayer_components::multi::traits::chain_at::ChainTypeAtComponent;
+use hermes_relayer_components::multi::traits::relay_at::RelayTypeAtComponent;
+use hermes_relayer_components::multi::types::index::{Index, Twindex};
 use hermes_test_components::driver::traits::channel_at::ChannelGetterAt;
-use hermes_test_components::driver::traits::types::birelay_at::BiRelayTypeAtComponent;
-use hermes_test_components::driver::traits::types::chain_at::ChainTypeAtComponent;
 use hermes_test_components::driver::traits::types::chain_driver_at::{
     ChainDriverGetterAt, ChainDriverTypeAtComponent,
 };
-use hermes_test_components::driver::traits::types::relay_at::RelayTypeAtComponent;
 use hermes_test_components::driver::traits::types::relay_driver_at::{
     RelayDriverGetterAt, RelayDriverTypeAtComponent,
 };
-use hermes_test_components::types::index::{Index, Twindex};
 use ibc_relayer_types::core::ics24_host::identifier::{ChannelId, ConnectionId, PortId};
 
 use crate::contexts::chain_driver::CosmosChainDriver;
