+++ crates/cosmos/cosmos-integration-tests/src/impls/test_driver/types.rs
@@ -2,12 +2,12 @@ use cgp_core::prelude::Async;
 use hermes_cosmos_relayer::contexts::birelay::CosmosBiRelay;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_cosmos_relayer::contexts::relay::CosmosRelay;
-use hermes_test_components::driver::traits::types::birelay_at::{
+use hermes_relayer_components::multi::traits::birelay_at::{
     HasBiRelayTypeAt, ProvideBiRelayTypeAt,
 };
-use hermes_test_components::driver::traits::types::chain_at::{HasChainTypeAt, ProvideChainTypeAt};
+use hermes_relayer_components::multi::traits::chain_at::{HasChainTypeAt, ProvideChainTypeAt};
+use hermes_relayer_components::multi::traits::relay_at::{HasRelayTypeAt, ProvideRelayTypeAt};
 use hermes_test_components::driver::traits::types::chain_driver_at::ProvideChainDriverTypeAt;
-use hermes_test_components::driver::traits::types::relay_at::{HasRelayTypeAt, ProvideRelayTypeAt};
 use hermes_test_components::driver::traits::types::relay_driver_at::ProvideRelayDriverTypeAt;
 
 use crate::contexts::chain_driver::CosmosChainDriver;
