+++ crates/relayer/relayer-components/src/lib.rs
@@ -10,5 +10,6 @@ pub mod build;
 pub mod chain;
 pub mod components;
 pub mod error;
+pub mod multi;
 pub mod relay;
 pub mod transaction;
