+++ crates/test/test-components/src/driver/traits/channel_at.rs
@@ -1,9 +1,8 @@
 use cgp_core::prelude::*;
 use hermes_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
 use hermes_relayer_components::chain::types::aliases::{ChannelIdOf, PortIdOf};
-
-use crate::driver::traits::types::chain_at::{ChainTypeAt, HasChainTypeAt};
-use crate::types::index::Twindex;
+use hermes_relayer_components::multi::traits::chain_at::{ChainTypeAt, HasChainTypeAt};
+use hermes_relayer_components::multi::types::index::Twindex;
 
 #[derive_component(ChannelGetterAtComponent, ChannelGetterAt<ChainDriver>)]
 pub trait HasChannelAt<const CHAIN: usize, const COUNTERPARTY: usize>:
