+++ crates/test/test-components/src/driver/traits/types/mod.rs
@@ -1,7 +1,4 @@
-pub mod birelay_at;
 pub mod builder_at;
-pub mod chain_at;
 pub mod chain_driver;
 pub mod chain_driver_at;
-pub mod relay_at;
 pub mod relay_driver_at;
