+++ crates/test/test-components/src/setup/impls/relay.rs
@@ -3,12 +3,12 @@ use hermes_relayer_components::build::traits::components::relay_from_chains_buil
 use hermes_relayer_components::build::traits::target::relay::{RelayAToBTarget, RelayBToATarget};
 use hermes_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
 use hermes_relayer_components::chain::types::aliases::ClientIdOf;
+use hermes_relayer_components::multi::traits::chain_at::ChainTypeAt;
+use hermes_relayer_components::multi::traits::relay_at::{HasRelayTypeAt, RelayTypeAt};
+use hermes_relayer_components::multi::types::index::Twindex;
 
-use crate::driver::traits::types::chain_at::ChainTypeAt;
-use crate::driver::traits::types::relay_at::{HasRelayTypeAt, RelayTypeAt};
 use crate::setup::traits::builder_at::HasBuilderAt;
 use crate::setup::traits::relay::RelaySetup;
-use crate::types::index::Twindex;
 
 pub struct SetupRelayWithBuilder;
