+++ crates/test/test-components/src/setup/traits/port_id_at.rs
@@ -1,9 +1,8 @@
 use cgp_core::prelude::*;
 use hermes_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
 use hermes_relayer_components::chain::types::aliases::PortIdOf;
-
-use crate::driver::traits::types::chain_at::{ChainTypeAt, HasChainTypeAt};
-use crate::types::index::Twindex;
+use hermes_relayer_components::multi::traits::chain_at::{ChainTypeAt, HasChainTypeAt};
+use hermes_relayer_components::multi::types::index::Twindex;
 
 #[derive_component(PortIdAtComponent, ProvidePortIdAt<Context>)]
 pub trait HasPortIdAt<const CHAIN: usize, const COUNTERPARTY: usize>:
